+++ python/triton/compiler/__init__.py
@@ -1,0 +1,2 @@
+from .code_generator import CompilationError
+from .compiler import CompiledKernel, compile
